refactor BaseJDBCResource => JDBCResourceManager
