+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/common/jdbc/text/JDBCTextBackendHandler.java
@@ -25,8 +25,6 @@ import io.shardingsphere.core.routing.StatementRoutingEngine;
 import io.shardingsphere.proxy.backend.common.ProxyMode;
 import io.shardingsphere.proxy.backend.common.jdbc.JDBCBackendHandler;
 import io.shardingsphere.proxy.backend.mysql.MySQLPacketQueryResult;
-import io.shardingsphere.proxy.backend.resource.ProxyJDBCResource;
-import io.shardingsphere.proxy.backend.resource.ProxyJDBCResourceFactory;
 import io.shardingsphere.proxy.config.RuleRegistry;
 import io.shardingsphere.proxy.transport.common.packet.DatabaseProtocolPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.command.CommandResponsePackets;
@@ -52,7 +50,7 @@ public final class JDBCTextBackendHandler extends JDBCBackendHandler {
     private final RuleRegistry ruleRegistry;
     
     public JDBCTextBackendHandler(final String sql, final DatabaseType databaseType) {
-        super(sql, ProxyJDBCResourceFactory.newResource());
+        super(sql);
         this.databaseType = databaseType;
         ruleRegistry = RuleRegistry.getInstance();
     }
@@ -67,9 +65,8 @@ public final class JDBCTextBackendHandler extends JDBCBackendHandler {
     @Override
     protected Statement prepareResource(final Connection connection, final String actualSQL, final SQLStatement sqlStatement) throws SQLException {
         Statement result = connection.createStatement();
-        ProxyJDBCResource proxyJDBCResource = (ProxyJDBCResource) getJdbcResource();
-        proxyJDBCResource.addConnection(connection);
-        proxyJDBCResource.addStatement(result);
+        getJdbcResourceManager().addConnection(connection);
+        getJdbcResourceManager().addStatement(result);
         return result;
     }
     
@@ -82,7 +79,7 @@ public final class JDBCTextBackendHandler extends JDBCBackendHandler {
     protected QueryResult newQueryResult(final CommandResponsePackets packet, final int index) {
         MySQLPacketQueryResult result = new MySQLPacketQueryResult(packet);
         if (ProxyMode.MEMORY_STRICTLY == ruleRegistry.getProxyMode()) {
-            result.setResultSet(getJdbcResource().getResultSets().get(index));
+            result.setResultSet(getJdbcResourceManager().getResultSets().get(index));
         } else {
             result.setResultList(getResultLists().get(index));
         }
